test(accordion): reorganize test files
