+++ packages/accordion/test/dev-mode.test.ts
@@ -1,0 +1,22 @@
+/*
+Copyright 2020 Adobe. All rights reserved.
+This file is licensed to you under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License. You may obtain a copy
+of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+OF ANY KIND, either express or implied. See the License for the specific language
+governing permissions and limitations under the License.
+*/
+
+import { fixture } from '@open-wc/testing';
+
+import { Accordion } from '@spectrum-web-components/accordion';
+
+import { Default } from '../stories/accordion.stories.js';
+import { testForLitDevWarnings } from '../../../test/testing-helpers.js';
+
+describe('Accordion - dev mode', () => {
+    testForLitDevWarnings(async () => await fixture<Accordion>(Default()));
+});
